[tf] Refactoring top-level security tests (#39453) (#39743)
This refactors all of the top-level security tests to use the
common echo deployment and the new echotest framework. Also moves
the authz tests back to the top-level, since it can now share
the same TestMain.

This is a roll-forward of the original PR.
Reverts commit 029d6d267710c0cc061072b1995c6f34929cd50d.