Disable full push on scale from 1->0->1 (#40866)
* Disable full push on scale from 1->0->1

Fixes https://github.com/istio/istio/issues/39652

This reverts https://github.com/istio/istio/pull/36882. At the time,
that PR was needed because EDS ServiceAccounts and CDS ServiceAccounts
were decoupled; Since https://github.com/istio/istio/pull/39133, this is
no longer true, and the fix in 36882 is not needed any longer.

This PR *removes* the test added in 36882 (since it tests low level
details that are not relevant anymore). It improves the existing
TestEndpointFlipFlops test -- while that test *would* have caught the
regression, it didn't actually set any service accounts so it was
missed. The update changes it to correctly detect the behavior (it now
fails without this PR, passes with it).

* add note

* fix note

* Add new SA case