gateway-api: update types to read v1beta1 (#41092)
* gateway-api: update types to read v1beta1

The upstream API is dropping support for the v1alpha2 types for
HTTPRoute, Gateway, and GatewayClass in the next release. So, unlike
Istio where we always read the alpha versions in Istiod, we must
switch to reading the beta types or users on v0.6.0 (next release of
API) will be broken.

This means that the minimum version of the API is v0.5.0; most users are
upgrading quickly since v0.4.0 was still an alpha API.

* add note

* fixes

* fix unit tests

* fix tests

* Tcproute is alpha

* Use TCP app Protocol

* goldens