wasm test: do not rely on dns resolution (#41667)
Slightly change a test case to not rely on dns resolution.

My system has a slow dns and this test fails because every attempt to
reach `https://dummyurl` takes between 3 and 5 seconds. The test case
timeout is 10s, which gives room to approximately 2 attempts, not 5 as
the test expects.

We could also increase the timeout, but I don't think it's the correct
approach.

Changing the test case URL to an invalid one gives us the same result
(i.e., testing a download failure) whithout relying on DNS resolution.