ambient: properly opt-out sidecar pods  (#43676)
* ambient: properly opt-out sidecar pods

Currently, we try to skip sidecar pods by looking for some labels on the
pod. This is not accurate, as it misses some cases where we get
sidecars. Instead, actually look to see if there is the sidecar, instead
of using labels.

This fixes a case where we may accidentally attempt to add ztunnel
redirection AND sidecars which is broken

* skip istio-system things as well

* gen

* More improvements

* lint

* remove container check