ServiceEntry IP allocation: Stable IP when used in multiple namespaces (#43867)
* test: test all elements in response array

* service registry IP allocation: Add test case triggering collision

* ServiceEntry IP allocation: Stable IP when used in multiple namespaces

On 1.17.1 we noticed a sudden BlackHoleCluster issue across many namespaces.

We see IP changes can be triggered when namespaces with ServiceEntries
with the same hostname is added. This is because the current
implementation uses a hash is created based on namespace and hostname to
allocate the IP:

	hashedServices := make([]*model.Service, maxIPs)
        // ...
			hash.Write([]byte(svc.Attributes.Namespace + "/" + svc.Hostname.String()))
                        hashedServices[firstHash] = svc

then, later, they are de-duplicated based on only the hostname:

	for _, svc := range hashedServices {
                // ...
		n := svc.Hostname.String()
		if v, ok := hnMap[n]; ok {
			log.Debugf("Reuse IP for domain %s", n)
			setAutoAllocatedIPs(svc, v)

This update introduces a small helper function makeServiceKey to make
sure the same key is used both places for stable allocation. A
regression test has also been added.

Fixes #43858