bench: drop obsolete test (#44032)
This tests the performance of the EnvoyFilter in the helm chart. This
could stillbe tested today, but its a lot harder since we need to
generate the chart, as the default generated yaml no longer includes
them. Since we expect most users to not have them anymore, remove the
benchmark entirely instead of dealing with the major complexity required
to fix up the benchmark