Fix ServiceEntry WorkloadInstanceHandler for label change in Pods (#42922)
* Fix ServiceEntry WorkloadInstanceHandler for label change in Pods

For a label change in WorkloadInstance, which results in a mismatch with a previously matching ServiceEntry, the whole update event is skipped. Which results in the WorkloadInstance not getting removed from the ServiceEntry.
This fix changes the behaviour for such cases to handle those updates similar to a delete event and clean up the WorkloadInstance from the list.

Fixes [#42921]

* use DeepEquals to avoid SubsetOf iteration

* gocritic linting fix

* go back to the first approach

* fix workloadInstanceDiff

* fix linting errors

* change test cases to table format

* fix testcase for new Event type

* fix Event Type

* go fmt

* remove wi diff