kube: drop MockClient (#44704)
* kube: drop MockClient

MockClient was a mix of usage where the FakeClient should have been
used, and overriding portforward/exec (which fake does not support).

This refactors to move most things to FakeClient, and make MockClient
embedd that and just overwrite the few fields we need.

This doesn't have huge wins on its own, but with another PR I have
locally it cuts about 10mb off the Pilot binary.

* lint