spiffe: fix handling of trust bundles with multiple keys (#44831)
In the existing implementation, we were overriding the `cert` while
iterating over doc.Keys. This commit fixes that.

Further, there was an unnecessary check for the existence of
`ret[trustDomain]`. We are iterating over a map with `trustDomain` as
keys, maps don't have duplicate keys, and so `ret[trustDomain]` would
never have had a key before we set it in this iteration.

Change-Id: I4132cd1edb18e415f5d1c228c04d60adaddc7445