Fix invalid XDS configuration for wildcard Ingress HTTP path (#44898)
* Fix invalid XDS configuration for wildcard Ingress path

Updates Ingress to VirtualService translation to not create a
HTTPRequestMatch when the URI is nil. The URI is nil when the path
is a wildcard or is empty and the pathType is nil or
implementationSpecific. This change prevents an Envoy failure.
Envoy regex fails when the path seperated prefix is empty or has a
trailing "/".

Signed-off-by: jaellio <jaellio@microsoft.com>

* Fix failing unit tests - handle sorting HTTPRoutes with no
HTTPRequestMatches

Signed-off-by: jaellio <jaellio@microsoft.com>

* Update golden converion tests for ingress. Remove empty match.

Signed-off-by: jaellio <jaellio@microsoft.com>

---------

Signed-off-by: jaellio <jaellio@microsoft.com>