cluster: cleanup ServiceInstances usage (#45868)
* cluster: cleanup ServiceInstances usage

We had a bit of a code smell earlier, where we create a fake struct
because the function signature expects it -- but we don't actually use
the data. This has led to bugs in the past.

I think its best to pass in only what we need. Here I make a new type
that is just a Service+Port+TargetPort, without the endpoint
information.

Part of https://github.com/istio/istio/issues/45241

* new name