logging: cleanup errdict usage (#46011)
For https://github.com/istio/istio/issues/44683

Partially reverts https://github.com/istio/pkg/pull/192

We current have an `errdict` mechanism, which is basically just a struct
of k/v pairs to insert into logs. We made this builtin to our logging
system. This adds a lot of complexity, and I don't think its needed.

Instead, just add a helper to the errdict to make it add labels to the
logger we want, and strip out all of the first-class support in
`pkg/log`.

This change moves us closer to being able to just swap `slog` in