Dualstack: fixing to not ignore link local addresses in proxy detection (#46016)
* Signed-off-by: abasitt <abdul.basit@rakuten.com>

This will fix issue #45335. This issues happens only if there are secondary interfaces in the POD.

We are ignoring linklocal addresses for proxy mode detection (Singlestack or dual). There was a previous fixed done in PR#37164 that hardcoded ignoring IPv4 linklocal address for EKS and on top of that users have to exlude 0.0.0.0 so the init consider don't redirect IPv4 traffic to the proxy.

Recently we have another fixed which is more flexible and cleaner by ignoring the excluded interface. E.g. If primary interface is IPv6 and secondary interface is IPv4, user can decide below.
- Install istio in singlestack mode and exclude IPv4 interface
- Install istio in dualstack mode and don't exclude any interface

* once interface is in the excluded list, exclude all IP addresses including multicast and loopback IP

* adding release notes