slice/maps: move to stdlib (#47469)
* slice/maps: move to stdlib

This moves our usage to the stdlib insted of exp. Exp is experimental,
so the intent is to have everyone move to stdlib eventually.

The big change here is that SortFunc is now returning an int instead of
bool. This is kind of annoying for most of our cases. To work around
this, I did:

* Made a new `SortBy` which covers 90% of cases
* Made an example of how to properly sort by multiple fields

And moved all usages to one of those two.

I also included the linter rule to ensure we always use the right
packages.

Finally, the change made a real behavioral change -- we were
unintentionally sorting backwards in istioctl output; so I updated the
test to align with the new correct output.

* lint

* bump go.mod, lint