iptables: refactor to avoid global variables (#47644)
* iptables: refactor to avoid global variables

Currently, we have a complete mess of configuration for the iptables
commands. This is driven a few issues:

* It was designed to be *byte for byte* compatible with the original
  Bash script years ago, so it has a ton of weird bash-isms.
* It is designed exclusively as a command line, but now its used beyond
  that (in CNI, etc)
* Uses lots of globals, and each config item ends up in 3-4 places.

This PR refactors this to simplify the configuration surface. The actual
goal of this is to make it safe to run the iptables setup multiple times
in a long running fashion; due to global variable usage this was
extremely hard to do so safely previously.

As part of this, a new package `pkg/flag` is introduced. This replaces
our minimal usage of viper with ~3 lines of code to allow setting a
flag's default value from an environment variable.

There should be no behavioral differences, except some of the `--help`
output is more clear (defaults shown)

* Copyright banner