Fix the custom injection on OpenShift (#47898)
Use of custom injection (i.e., providing a custom `istio-proxy`
container) is broken on OpenShift.

This is because OpenShift sets the `RunAs` field to be the a
pseudo-random value that varies by namespace. We rely on the `RunAs`
field to be a well-known value, and it's calculated upon injection. We
should stick with that value. `iptables` rules were created based on
that value.

Even if the user provides another value (i.e., it's not set
by OpenShift, but b the user), that's wrong. We should ignore what's in
the original pod and use the calculated one, for these two fields only.