kclient: do not panic on handlers after stopped (#49184)
Currently, if we add a handler after the informer has stopped, it will
crash. This is not great, though it rarely happens.

Better handling of this is to just not add the handler.

Note this cannot be handled by clients, since there would be TOCTOU race
conditions (and we don't expose HasStopped anyways)