Separate CNIConfig and CNIUsageConfig (see #49290) (#49319)
We currently use the same protobuf message for both cni and istio_cni
even though istio_cni supports just a tiny subset of the fields in cni.
This PR is meant to fix that and prepare the ground for moving/renaming
of istio_cni in the future.

See https://github.com/istio/istio/issues/49290