Increase NOFILE limit for Envoy (#49770)
* Increase NOFILE limit for Envoy

Follow up to
https://github.com/istio/istio/pull/49761#issuecomment-1981362376.

Before, we hack around limits in bash to make VMs work fine. Majority of
k8s platforms do not need this, since they always have
softlimit==hardlimit and a reasonable value, but VMs usually have 1024.

With this fix, we ensure from the agent that Envoy always has the
highest limit, so we can remove all our bash hacks.

If we do not do this, we need to revert
https://github.com/istio/istio/pull/49761

* lint