helm upgrade tests: Upgrade from latest patch release (#50827)
Instead of upgrading from `.0`. For instance, if latest `N-1` release is
1.21.2, the test deploys Istio 1.21.2 (not 1.21.0 as it does today) and
then performs an upgrade to the current dev version.

For this we make use of the `~` in the version parameter of `helm install`.
Instead of invoking e.g., `helm install  --version 1.21.0`, we use `helm
install --version ~1.21.0`. This instructs helm to install the latest
patch release in 1.21.x: https://helm.sh/docs/chart_best_practices/dependencies/#versions

While on it, removed the workaround for 1.17. It's EOL.