validation: avoid redundant api server requests on each startup (#51237)
Currently, EVERY Istiod instance makes dry-run requests to the API
server to check if the webhook is ready to move from Ignore->Failure.

This PR changes it so we only do this check if the policy is currently
Ignore. This means in the common case, only on first install will we do
this; any new Istiod starting will do nothing (since the policy would
already be Failure)